Changed strcmpi() to lstrcmpi() u_long cast first argument in min() at http.cpp Corrected getdefaultwd() call in filemanager.cpp
To fix broken build on linux


git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@289 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/http.cpp
@@ -920,7 +920,7 @@ int controlHTTPConnection(LPCONNECTION a,char *b1,char *b2,int bs1,int bs2,u_lon
 					{				
 						u_long tr=min(content_len-total_nbr,td.buffersize2);
 						err=td.connection->socket.recv(td.buffer2,tr, 0);
-						td.inputData.writeToFile(td.buffer2,min(err,content_len-fs),&nbw);	
+						td.inputData.writeToFile(td.buffer2,min((u_long)err, (content_len-fs)),&nbw);	
 						total_nbr+=nbw;
 						timeout=clock();
 						break;
